+++ packages/core-js/internals/a-function.js
@@ -1,5 +1,4 @@
 module.exports = function (it) {
-  if (typeof it != 'function') {
-    throw TypeError(String(it) + ' is not a function');
-  } return it;
+  if (typeof it === 'function') return it;
+  throw TypeError(String(it) + ' is not a function');
 };